requirements.txt


fastapi
PyPDF2
python-pptx
sqlalchemy
pydantic
passlib
python-jose
python-dotenv
requests
pymysql
scikit-image

opencv-python
imutils
img2pdf
python-multipart